pyproject.toml
[project]
name = "pyautomouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pynput>=1.7.6",
    "colorama>=0.4.6",
    "rumps>=0.4.0",
    "pyobjc-framework-Cocoa>=9.0",
    "Pillow>=10.0.0"
]

[build-system]
requires = ["setuptools>=69.0.0", "py2app>=0.28.6"]
build-backend = "setuptools.build_meta"
